requirements.txt
# The Rational Decision Engine — Groq + DeepSeek only
openai>=1.0.0
yfinance>=0.2.40
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
